feat(medication): Update medication attributes and add facility query for dispensing workflow
